+++ web/app/themes/brocooly/config/views.php
@@ -13,10 +13,11 @@ return [
 	 * View namespaces
 	 * --------------------------------------------------------------------------
 	 *
-	 * You may add extra namespace into your view files
+	 * When you refer path to a template from within twig file, you’ll use the relative path from the main `resources/views` directory at the root of the project. If you need you may specify here a shortcut for some 'long' path (nested directories) and use with `@` prefix
 	 *
 	 * @example
 	 * ```
+	 * 'components' => 'components' means `@components` will correspond `components` directory
 	 * {% include '@components/path/to/view.twig' %}
 	 * ```
 	 */
@@ -38,7 +39,7 @@ return [
 	 * Templates
 	 * --------------------------------------------------------------------------
 	 *
-	 * Register custom templates
+	 * Register custom templates. Pass special templates classes here
 	 *
 	 * @link https://developer.wordpress.org/themes/template-files-section/page-template-files/
 	 */
@@ -60,7 +61,8 @@ return [
 	 *
 	 * All your directories for view files should bу listed here.
 	 * Relative path to a theme root folder.
-	 * ! MUST be string in opposite to Timber documentation
+	 *
+	 * @var string|array
 	 */
 	'views'                => 'resources/views',
 
@@ -72,6 +74,9 @@ return [
 	 * When you are getting all posts you may set posts_per_page query parameter as -1.
 	 * It is not recommended way so we're set it 100.
 	 *
+	 * This settings applied for `Post::all()` etc.
+	 *
+	 * @var int
 	 * @since 0.10.2
 	 */
 	'limit'                => 100,
@@ -84,6 +89,7 @@ return [
 	 * In case you miss to include some view file or unexpected output occurs
 	 * this file will be rendered instead of 'undefined template' error
 	 *
+	 * @var string
 	 * @since 0.8.5
 	 */
 	'default'              => 'callback.twig',
@@ -97,6 +103,7 @@ return [
 	 * It is available only if WP_ENV === 'production'.
 	 * Set `false` to disable
 	 *
+	 * @var string
 	 * @link https://timber.github.io/docs/guides/performance/
 	 */
 	'cache'                => BROCOOLY_THEME_PATH . '/storage/cache/',
@@ -108,6 +115,8 @@ return [
 	 *
 	 * Set time in seconds cache files will be expired.
 	 * Default is 600 (10 minutes)
+	 *
+	 * @var int|bool
 	 */
 	'expire'               => 600, // seconds.
 
@@ -117,6 +126,7 @@ return [
 	 * --------------------------------------------------------------------------
 	 *
 	 * Define is maintenance mode enabled or not
+	 * If it use maintenance template on front of your site.
 	 *
 	 * @since 0.14.0
 	 */