+++ web/app/themes/brocooly/src/Containers/BlogSection/Tasks/GetBlogTitleTask.php
@@ -4,12 +4,13 @@ declare(strict_types=1);
 
 namespace Theme\Containers\BlogSection\Tasks;
 
+use Brocooly\Support\Facades\Option;
+
 class GetBlogTitleTask
 {
 	public function run() {
-		$title = get_option( 'page_for_posts' ) ?
-				get_the_title( get_option( 'page_for_posts' ) ) :
-				esc_html__( 'Blog', 'brocooly' );
+		$blogPageId = Option::get( 'page_for_posts' );
+		$title      = $blogPageId ? get_the_title( $blogPageId ) : esc_html__( 'Blog', 'brocooly' );
 		return $title;
 	}
 }